[boursedirect] Use the response declared encoding and ignore the document declared charset
Despite the webpage meta charset declaring to be UTF-8, some pages
content is not UTF-8 valid content.
Ex.: iso 8859 é (0xe9).

In fact, most of the content is simple uppercase ascii, but sometimes
accounts name could be missing and be replaced by the string:
"Compte non trouvé".

The encoding in the response header content-type looks like to be
correct.