[boursedirect] Fixes a crash for stocks without an isin code
There could be stock without an available isin.
In that case, the parsed url will have an empty isin arg and be like:
<li><a href="/priv/mobilalert.php?isin=&page=Events">

We change the regexp to consider the "empty" isin a valid value.
And in the caller, the is_isin_valid will be False, and so inv.code will
keep the NotAvailable value for this entry.