[bp] handle deferred cards that have empty summary history
Some deferred cards do not have transaction history and can even make
the website crash if we try to access it. It seems to be linked to
whether the card is from a third party account or not. In any case
retrieve the card details via an API call, not via the transaction
summary page.