[cmso] Fetch loans on pro website
The page is very similar to the loans page on the arkeabanqueprivee pro
website, but it's a list instead of a table, so all xpaths are
different.

Like on arkeabanqueprivee pro, loans have no number and no ID on the
website; only a label which isn't unique. We have no choice but to use
the "index" which is an unstable JS generated value.