[s2e] Replace stay_or_go call to simple go
The s2e website is stateful, and a lot of different content can match
the same url, so it is safer to always go back to the inital page.

By only using a go, we can be sure that the good page will be loaded in
the correct state before trying to use any function of the page.