[s2e] Detect when incorrect page content is returned because of of disconnection
Sometimes, instead of the expected content, an empty page with just a
META redirection is returned when it looks like that the user was
disconnected.

Ex.:
<!DOCTYPE html><html><head><META http-equiv="refresh" content="0;
URL=/portal/salarie-axa/operations/consulteroperations"></head><body></body></html>

A further request would encounter 302 redirections to the login page.

In such a case, raise a "Loggedout" exception as previously it was just
silently ignored and considered as a normal page.

Note: check_disconnected as to be called manually in iter_history, as a
simple "open" (compared to a "location") will not call the "on_load"
function of the page.