[s2e] show_more should be ok if we already have the current number of items per page
show_more() is called to select the number of items per page.
It will return False if the selector was not found, and True if it was
found and successfully changed to the request value.

But, currently, if we found the selector but it was already on the
correct value, we would return nothing, that would be equivalent to
False for the "if" test that is used in the browser.py.

This issue associated with a previous issue when "stay_or_go" was used,
led to having the history entries returned only for a single time, for
single account, but not for subsequent calls or different accounts.