[s2e] Only go back to history list from history details once per page
As the website is  stateful, after being on a "History list", if we do a
request for the details of one entry of the list, we will have to send
the "going back" request to have the "History list" page again.

But, it is not needed to "go back" to see other "details" pages.
So, this change improve the performance by only going back once before
trying to go to the next page of a History list page.

In addition, this might fix some nasty issues as the website is
"unstable" and could easily have timeout issues. And in that case,
missing any step could easily break next requests with strange errors.