[lcl] Loan revolving: handle missing loan rate and amount column
Setup right column for balance for loan revolving, we want to display
the used amount, which is the amount that need to be refund.
Total amount for loan revolving match the max amount authorized by the
bank.
For this kind of loans, LCL doesn't display any rate, handle this case.