[caissedepargne] Handle deletion for duplicated cookies
Sometimes we have cookies with duplicated names in the CookieJar. This can
happen when we have cookies with the same name but with different
domains or paths.

We can't access or delete these cookies as we would normally because it
would raise a CookieConflictError. We have to use the
`remove_cookie_by_name()` function for that particular case.