[amazon] Prevent sms spam and remove restricted URL to not trigger sms sending
Previously we were sending 1 sms every minutes until the connection is approved
by the user. Moreover there was a restricted URL we were trying to access which
led us to trigger the in-app validation which sends an sms/e-mail/in-app
notification to approve the connection even if we were already logged in.