[carrefourbanque] Fix login
The first time we login, we get a cookie 'SSESSION...'.
If this cookie is present, when we try to go to the LoginPage we will be
redirected to the HomePage.

For the next session, because locate_browser() doesn't do anything, do_login() will be called
but if the cookie 'SSESSION...' is valid we'll be redirected to the HomePage.

This results in calling LoginPage's enter_login() method, even though
we're on the HomePage, and it causes an AttributeError.

To solve this I removed locate_browser() overriding.