[carrefourbanque] Fix login and blackbox
If we try to access the LoginPage when we're already logged in, we are
redirected to the HomePage.

Overriding `locate_browser` is also necessary to make the blackbox work.
If we just `pass`, we would never get to a `LoggedPage`. So a solution
in which we check if we're on `HomePage` in `do_login` is not
sufficient.