browser.filters.base: make @debug more lazy
The @debug decorator is put on many filters and computes a lot of stuff
then verbosely logs it with a very low logging level (lower than DEBUG),
that by default isn't displayed.
Instead of computing so much stuff for nothing (logging level not low
enough) in the general case, only do it if we know we need it.

Also, it avoid having to skip a lot of useless code when running inside
a debugger.