[caissedepargne] screen out wrong transaction rdates
Sometimes transactions of checking accounts have a deceiving label that
seemingly contain an rdate whereas it is only some kind of
transaction/product identifier. The problem is we still parse this
identifier as an rdate:

ex: label "CB <the_store_name> 231001" will be matched by regex
r'^CB (?P<text>.*?) (?P<dd>\d{2})(?P<mm>\d{2})(?P<yy>\d{2})\b' (with
ignorecase flag) and the rdate will be set as 2001-10-23 even if 231001
is not a date.

It doesn't seem possible to improve the way we parse the label (same
transaction layout as other labels that have a correct rdate), except by
manually excluding these wrong labels, which is not appropriate either.

Solution: one way to distinguish false rdates is to check if the rdate is
close enough to the transaction date, directly in the module.

NB: these transactions are also fetched for the corresponding deferred
card and in that case they have the correct rdate (the transaction
labels in the card page is more comprehensive and contains the rdate)