[amazon] handle specific cases at login
Previously we were requesting the history page to know if we were logged or not
but it seems that sometimes the history page can pass through the login page but
it doesn't mean we are connected. Now we'll also be properly redirected to the
page we were trying to reach before do_login and the reset password page is
properly handled.