[amazon] raise AuthMethodNotImplemented for specific 2FA
For some accounts after the login page we are redirected to the new-otp page
which asks us to send a notification but after sending the code we are
redirected to the new-otp page again and forever and we can't manually call
the url where we can validate the otp since it's generated server side. At
the moment this case can't be reproduced even with some hacks, so this is a
temporary solution to be able to login for the accounts involved.