[caissedepargne] handle instabilities during authentication
Instabilities tend to happen when the main website of Caisse d'Epargne
is redirecting to their auth service (icgauth.caissedepargne) to
authenticate the user with an SAML protocol. The auth service fails to
initiate the user auth (failure during submission of the SAML request
data).

Monitoring shows the instabilities seem to last a few hours, retrying
the authentication process doesn't seem to work well as long as
the auth service is unstable).