[caissedepargne] fix how we build urls for millevie insurances
So far millevie's urls are built thanks to accounts' labels. But for some
reason the labels are not retrieved the same way as they were before.
Consequently the millevie's urls are not complete and lead to an error
page.

Instead, use the accounts' number (only for natixis' accounts, which
include millevie life insurances).