[hsbc] Handle relogin after otp validation
In some scenarios, relogin are triggered. If it happens inside
`handle_otp`, the otp value stored in config will not be cleared.

The consequence is that we will try to validate the otp once again, even
though we might not even be on the right page.

To fix that we reset the otp value to its default value manually.