[bnp] Check for otp and special cases after the confirmation step of transfers
After the "enregistrerVirement" request, an "ok" response does not
always mean that the transfer initiation is already completed.

In some cases, an additional step of sms otp or mobile app validation is
needed.

In some other cases, the transfer is not really registered and will not
be executed as the amount is too big or it exceeds daily limits.
In such a case, on the bank website, an email would be sent to the
client bank agency to ask for a manual approval of the operation.

As we don't support any of these operations yet, we just raise errors
when encountering them, to avoid "false" confirmations.