[fortuneo] Handle post-otp state_expired
During the login, the user has to validate an otp. After sending the otp,
the call to `check_and_handle_action_needed()` may raise an ActioNeeded.
In that case the otp config_key will not get cleared.

If the user takes more time than the STATE_DURATION (5mn at the time of this commit)
to do what he needs to do, the state will expire and get cleared. When the PSU will try
to update its connection, it will still call `handle_sms()` method but
the `sms_form` will be None and the otp value might not even be valid
anymore.

To fix that, we clear the otp config_key manually.