[fortuneo] Avoid handle_sms() after an ActionNeeded
An action needed can happen during the handle_sms,
but self.sms_form will have been re-initiated to None while the user resolve it,
and the OTP will already been submitted and accepted by the server.
So, to avoid running handle_sms a second time, we check if self.sms_form is present;
when not, we fall back to init_login, where the SCA won't be triggered.
This avoids a "'NoneType' object does not support item assignment"
error.