[hsbc] Avoid false positive error messages
commit 8087d2f6dea023946894032a8fde7c1bac6438ce added the handling of
more error cases and a check for error messages at the `end_login()`
step to fix `FormNotFound` errors.

At this step, error messages and the message upon a succesful login are
located under the same XPath. So, we might raise false login errors.

To avoid false positives, we check whether we are logged in or not before handling
errors.