[bred] Simplify get_accounts_list and ensure parent matching is done with accounts of all universes
Sometimes, the life insurance account will be found in the same univers
as the parent account and sometimes it is not the case.
In addition to that, the iteration order of the univers for the
get_accounts was having an impact on whether the "parent" could be
filled or not for a life insurance account.

Now, we do the matching only after all the accounts have been retrieved.