[ganpatrimoine] Handle `Certificat mutualiste`
`Certificat mutualiste` used to be a category. Now these accounts are in
the `Epargne` category, but their details page is not the same as other
`Epargne` accounts, so a specific case is needed to handle them.

The issue appeared with the new version of the website in 2019 but
these accounts were classified as insurances by the website and they
were not fetched at all.