[bforbank] Fix URL to investments space
The previous way to fetch the URL of market accounts now returns a
generic URL which results in an error. The correct URL is in a different
attribute.

I tested that the new URL does lead us to the netfinca space, however
I could not test the relogin after leaving it because of the OTP.

Closes: 34197@zendesk, 34336@zendesk, 34355@zendesk, 34368@zendesk
Closes: 2715@sibi