[creditmutuel] Raise warning instead of error for cards in tech error
When the page which contains the "functions" of a card is unavailable
due to a technical error, we log a warning instead of raising
BrowserUnavailable for the whole connection. We have most info for the
card (it's the same behavior as when the user is not authorized to
access the page).

Closes: 34652@zendesk