[axabanque] Handle undownloadable documents and downloads with a URL
Some documents are not downloadable with the usual URL (which uses the
`download_id` as a parameter) but with a specific URL which can be
fetched from the detail page.

Other documents are not downloadable and the link leads to a page with a
form to be sent the document by mail (not email, mail).

Finally, the URL for contracts has changed, and some documents have no
date.