[s2e] Fixes wrong quantities for investments and pockets
The "diff" value was reported as "quantity".

On s2e websites, the investments pages can show the list of investments
can be seen with 2 different tables/views. One that displays
"valuations" and the other one that displays quantities.

We go from one view to another through a input switch.
Before, it was enough to set a "visualisationMontant" or
"valorisationMontant" boolean to have the expected view.
Now, it is also needed to add a specific input to "activate" the view
that we want.

As this was missing, we were also using the "valuation" view, and so
trying to retrieve the last column anyway for quantities (even if it was
"diff" in this case).