[s2e] hsbc,erehsbc: Rework and fixes to go to fund info pages and get amf codes
For some funds, we get a link that directs directly to the fund info
page as they are "private"/"closed" funds.

For common public founds, the "direct link" will not work, but we can
search for them in hsbc fund search engine.

And, for some special funds, sometimes there is a link to a
"videos-pedagogiques" page that has a link to the fund info page for all
classes of the share.