[s2e] Fixes "objects with the same ID" issue
Since the s2e website was modified, accounts "name and id" list is on a
different page than the accounts "summary" page.

And, sometimes, 2 different "accounts" can have the same name (Ex.:
PEG).
In that case, there is absolutely no info on the summary page that could
allow to differentiate between the 2 accounts listed with the same name
to find the correct id.

To resolve this issue, we assume here that the accounts will always be
listed in the same order on both pages. And so, in case of "conflict",
we attribute the id for a name in the same order as they are listed on
the "id list" page.