[bforbank] fix IBAN and card page with account._url_code
Recently some URLs have changed their account id param.
Instead of the account id, they needed the new _url_code
attribute. Not using this one leads to error 500 with a
maintenance message that is in fact deceptive.

Since the implementation of the new URLs, IBAN were not retrieved anymore.
That is because the RIB request also need to use _url_code.
This commit fixes this, as well as simplifying and cleaning the way
IBAN is retrieved.

_url_code must also now be used to access card_page
or we also end up on a error 500.