[cmso] Fix accounts matching when fetching the list of transfer recipients
The label format changed with !790, leading to a difference in the hash
used to match accounts between what is fetched by
AccountsPage.iter_accounts and RecipientsListPage.iter_int_recipients.

With this fix AccountsPage.iter_accounts no longer relies on the label
field to compute the account matching hash, allowing to change the label
format without breaking the fetching of transfer recipients.