woob.capabilities.bank.wealth: comment on Pocket for bad conditions
The only correct conditions are AVAILABLE, DATE and RETIREMENT.
The other conditions were mistakenly added.
They do exist but they are secondary, they are always in conjuction with one
of the 3 main ones, they are not mutually exclusive.
If condition is RETIREMENT, the now-deprecated elements can still apply
earlier than retirement, and you will never see as sole condition one of the
deprecated conditions.