woob.browser.TwoFactorBrowser: handle twofa date more consistently
Previously the attribute twofa_logged_date was kept as string, except:
- in the first synchronization
- during the call to get_expire() where it is converted back to
datetime

It is also not cleared each time,

In addition to being inconsistent (we expect the same type for the attribute
twofa_logged_date at any moment of the connection's life), it makes
the use of the twofa_logged_date less convenient because we then need to parse
the date back from string to datetime anytime before it can be used. So
the twofa date is now made available as a datetime in
self.twofa_logged_date, not as a string.