[myfoncia] Rework the module
The website has changed and moreover the login, which is a perfect
illustration of being a pain in the ass.
Myfoncia use a JS library called FingerprintJS. It compute an unique ID
per visiter based on some parameters such as the navigator and the
hardware on which is running. It's also capable of detecting whether or
not you're a bot.