tests: use pytest instead of nose
Warning: the reported coverage is much lower than before. But when reading
the previous reports, it seems the coverage was overestimated by large,
mostly because old report did not consider all files in woob dir.

Also for some reason pytest's JUnit XML uses `modules.<name>.<class>`
with the `modules.` prefix so we take it into account in testing grid.