woob.tools.date: Use tz.tzutc instead of gettz
gettz() was ok, but tzutc() should be marginally faster and consistent
with the rest of the code.

Also, `tzutc()` was already available in old versions of dateutil.
But, it should be a lot faster with versions >=2.7.0 as it will use a
singleton to always return a single preloaded object (ie tz.UTC)