[n26] rework the module
Rework module using the TwoFactorBrowser, in order to make
the module easier to debug.

What's new:
- The AppValidation 2FA method was implemented as it is more robust
as N26 has stated in their documentation. This method is used as a
default and we fallback to OTP by SMS if the user hasn't paired its
device.
- Added Pages instead of doing everything in Number26Browser
- Replaced `request` method by `build_request` as
it was only used to set the headers
- flake8 compatible

Fixes:
- Using an access token generated using the refresh token
gives access to 90 days of transactions only.
- Too many requests error sometimes don't return a json body.
- Clear access_token on 401 errors