[caissedepargne] properly set auth 'typ_act'
It was done wrongly before, if a non-interactive session
was opened while the 2FA validity had expired and if during
the `STATE_DURATION` time window you would start an interactive
session.

The "sso" typ_act would be used because the state wouldn't be expired.
To avoid such clumsy situations, we simply check if the 2FA validity has
expired right before setting the typ_act value.