[boursorama] Fixed ValueError for subscriptions label
For subscriptions, label divs are formed like "elem1 - elem2 (-optional
elem3)". We fetch obj_id and obj_label from it. The regex was looking
for both possible forms of id ("1234********5678" and "12345678912") but
sometimes, there are also digits in the obj_label like "Compte Epargne
2020". So the regex has been modified to look for at least 11 chars,
which is the minimum length of id's.