[creditmutuel] Redefine states instead of extend
The 'logged' attribute that is added to the StatesMixin __states__ is
causing problems in other modules.

'logged' in 'creditmutuel' is a simple attribute, but it can be  a
property in other modules. That causes a problem during the state loading.

We should use `self.__states__ = self.__states__ + [...]`.
We can also delete logged since it is not used anywhere.