[amundi] New login and changes for the parsing
Amundi's authorization token is now obtained on a new route called
virtualkeyboard (there's no VK for the moment). We must do a GET on a
specific route before that with a UUID that must be kept for the
virtualkeyboard page. BaseURL for Amundi Epargne Salariale has changed,
as well as Amundi TC and Amundi CA.
There's also a captcha that is now handled (Google reCAPTCHA v2). Some
of the JSON we use for parsing have a slightly different structure.
The accounts on the website are all in ActionNeeded (the user must
confirm his email), but fortunately, we can still scrap what we need and
ignore that for the moment.