[amundi] Set token header only on needed routes
amundi module gets a lot of data from other websites for investments.
Some of the websites were not allowing us to access them because the
authorization token header for browsing on amundi was always set. Now,
the token is only set on for the pages needing it on amundi website
only.