[axabanque] handle profile page unavailable
The profile page is unavailable on the old website.
For now we handle it by raising a BrowserUnavailable.
But this page is only used for getting the profile
and setting the accounts' ownership.
It's a shame to stop whole parsing of the website just for that.

Solution: not returning any profile and ownership
if the page is unavailable.

The connections using the new website are not impacted.
The bank will eventually fix the page.