[axabanque] better handle profile page unavailable
While trying to reach profile page, we can get
an error 500 on the old website.
Since it can happen when coming from any other Page class,
depending on the order of navigation,
we don't use a message checking method inside a specific page,
but build a local HTMLPage from here.