[bp] Detect when accounts are not authorized for transfers
Sometimes, some standard accounts are not allowed to do transfers.
It will be the case if the user account is in special situations like
being blocked or having the certicode+ that is not well setup anymore.

In that case, trying a transfer would result in a "Recipient not found"
error, as trying to access the transfer page will result in a
maintenance error.

To solve this issue, we will detect early, for each account, if they are
allowed to be transfer sources. This is done by checking the existence
of the "Virements" button on account pages.