[bp] Fixes the check for transfers from livrets to forbidden recipients
For "livret" accounts, transfers are only allowed to the user main
account. It is a legal requirement.

If such an operation is still attempted by the user, La Banque Postale
will give an explicit message to the user.

The retrieval for this message was broken due to site changes:
- the html document meta is lying about the encoding but the
content-type in the response headers is correct.
(Document is in fact in iso-8859-1 instead of utf-8)
- divs have 2 different ids suffixes ("inter", "intra") depending on the
type of the forbidden recipient (internal, external).