[ganpatrimoine] Detecting some cases where the sms sca will be needed
Not sure why, but when a SMS otp will be needed, in some cases, we are
facing an error page with the "non supported browser" error.
There, there is a link to go on with the sca process.
We have to stop there, because, otherwise, the SMS will be sent directly
if following the link.