[ganpatrimoine] Detect when website is in maintenance at access point
Sometimes, we are encountering a maintenance page, directly at the base
url: espaceclient.%(website).fr
And it is the page where we are supposed to get the login page.

Before, in addition with not detecting the website unavailability, we
were crashing in the following part of the code, as we were trying to
extract a "website" field from the url query parameters that would not
exist.
I did not change this part of the code that was crashing, as I was not
able to be in case where the child module would to use it, so I'm unsure
of that it is supposed to do exactly.