[ganpatrimoine] Improve the handling of /wps/myportal sub-site pages
There is a /wps/myportal sub-website inside the mainsite that is
dedicated to handle bank related accounts.

This change improve the handling of the pages of this mini website.
Currently, only the iter_history and iter_card_history functions using
this mini website are used by the ganpatrimoine browser to retrieve the
history of a few types of account.
This could be improved, but at the moment, all the added pages are to be
used by child modules that can only have access to the mini website and
no other area of the ganpatrimoine "espaceclient" website.