[bforbank] Fix life insurance history
Prior to this commit, trying to retrieve the history of a life
insurance resulted in an error 500. It seems the bank website has
changed its parameter `id_account` in the URL by a given code,
loaded when the accounts are retrieved.

The previous endpoint:

`/espace-client/assuranceVie/{id_account}`

is now:

`/espace-client/assuranceVie/{url_code}`

Fixing this allow to retrieve once again the history of a life
insurance account.