[s2e] Handle redirects during login
Sometimes we get redirected to `/portal/setcontrolcookie`
which returns a 404, that needs to be handled. A retry do the trick.

We also handle a new url: `/pagesdedelestage/BNPP/index.html`
which say that the server has a big load and is sometimes not
responding. Also a retry does the trick.

We also increase the number of tries of the retry, because there are
more things that could cause it and two tries may not be enough anymore.