[transatplan] Handle two accounts with multi plans
In some rare cases we can find accounts that contains two plans
(maybe more) with the same ID (pockets + stock options).

In this case we ignore the second plan and concatenate the
two plans to calculate the account balance.

We also handle tab navigation to get to the pockets page.