[orange] Retrieve rechargeable subscriptions
People with rechargeable mobile subscription did
not have the possibility to retrieve their bills.

This commit fixes that behaviour and retrieves the
concerned subscriptions and bills. However, unless
the user is an account manager or the CEO of his
company, he will not have the possibility to
download his bills; he would then only have access
to meta data.