woob/browser/elements: Add ItemElementRerootMixin
While woob works well with scraping information from websites
serving HTML, there are some limitations to implementing
support for JSON APIs using it; this commit aims at fixing
one of them.

HTML websites usually present collections and items differently,
the first usually being a table, a list or flexboxes, and the
second usually being full page with information scattered around
the header and on different sections in the page. Idiomatic woob
programming supports this by defining ItemElement subclasses
directly in HTMLPage subclasses for reading the elements in a
way that is appropriate to the structure of the page.

However, APIs tend to use a "one object, one format" approach,
whether it be in collections, in the single page, in search
interfaces, or other contexts. One such example could be a
payment API, where when created, the following is returned:

	{"createdPayment": {"paymentId": 1, ...}}

When payments are queried, the results look like this:

	{"payments": [{"paymentId": 1, ...}, {...}]}

And when a single payment is queried, the result looks like
this:

	{"payment": {"paymentId": 1, ...}}

The same fields are defined, but since they are at three
different places, using elements for such a problem yields
the following result:

* The first payment is defined as an ItemElement where every
  Dict selector is prefixed by "createdPayment/".
* The second page defines a ListElement with the item_xpath
  set to "payments", with an ItemElement where elements are
  unprefixed.
* The third payment is defined as an ItemElement, where every
  Dict selector is prefixed by "payment/".

This is widely impractical and violates the DRY principle.
In order to fix this, this commit adds an item_xpath property
to ItemElement, so that an ItemElement for a payment can be
defined with no prefix, and the first and third cases can
inherit from the ItemElement with the ItemElementRerootMixin
while defining reroot_xpath to "createdPayment" and "payment"
respectively.