[s2e] Reduce number of requests
Currently, for each transaction we make a request to get the list of
investments related to it. These requests make up a huge percentage of
the total requests we make during synchros. If we stop doing this it
will drastically reduce the load on s2e servers.

This also fix a navigation problem, currently we don't press the return
button after going to the details page. This causes us to keep getting
the same html for almost all transactions.

- current iter_history:
	- for each transaction, we get the list of investments
	- the amount = sum(investments.valuations)
	- no return to previous page

- next iter_history:
	- the amount = net_amount + net_employer_contribution_amount
	- we drop the transaction.investments attribute
	- return after visiting details page

Note:
      - this will only impact "monoaccount" connections and "monoaccount"
        spaces, in multispaces connections.
      - for connections with multiple accounts, we will still need
        to visit the details page, to know which transaction is
        related to which account.