[cragr] Fix PACA matching
An end user had a Prêt à Consommer Automobile (PACA) matched
as a loan, which made the browser get transactions for this
loan. This resulted in the site giving the transactions of
the main checking account for the account, thus making
duplicates for both accounts.

This commit fixes the matching of 'PACA' to consumer credit
instead of loan, fixing this behaviour.

Closes: https://gitlab.com/woob/woob/-/issues/437