[cmes] fix iter_investments
xpath has recently been fixed for investments list
which last line is not an invest.
But this line is not always present and in this case
we skip the last investment.

For skip this line without skip last investment
we can skip the yield if any raw have been found