woob.browser.pages: don't resolve entities in XMLParser
XML entities are known to be subject to potential attacks, among which
quadratic-blowup or file expansion. As a side note, lxml is protected
by default against network-access or billion-laughs.
This can be safer in the event woob is used to scrape untrusted sites, which
is not the case when using regular modules.