[erehsbc] Rework login
S2E children have all the exact same login except erehsbc which has now
a whole new login. The login is a javascript app so we need to mimic the
website as close as we can by encrypting the password and doing many
posts on the same URL to get different JSON callback responses. There's
an OTP because the website asks for the user to enter the code he has
received by email if the device connecting to the account is not a
trusted one. Everything after the login is left untouched.