[oney] Handle internal error
Here we handle an error that happen on the oney website
caused by difficulty to communicate between the SCA server
and the Oney server. It happend randomly and can last few
hours, in some case where the problem persist you have to
delete your terminal in the app and then reinstall
the app to resolve it. If it still persist then you have
yo contact oney support.
This error only display a generic error message on the UI