[bnp] Handle new redirection process for Hello Bank
BNP's login works by doing a POST for credentials, then it does a few
redirections, then it stops and we must use a redirection URL manually
to keep the redirection process going to the end. Hello Bank's login was
working like that too until now. It does all the process of redirection
alone after the POST with credentials. But still, we need - as it is the
case with BNP - to handle the redirection flow because an OTP and some
potential error responses can be found here. Also removed all the
"is_old" stuff for both logins because it was useful when BNP updated
their website a few months ago and some instability made them use the
old login sometimes. Searched for traces of it, can't find any use case
of these old URLs.