[caissedepargne] Anticipate JS file name change
Last week, banquepopulaire module fully crashed because we need to get a
JS file during the login process and the name of that file had changed.
This change only affected child module banquepopulaire but it's going to
be effective on caissedepargne too from 11/05/2022. To avoid the module
fully crashing like banquepopulaire, we are going to handle both cases of JS
file names before they change it.