[boursorama] Handle new ActionNeeded case
ErrorPage was only used to detect cases where the user was temporarily
blocked. It can now display an information message saying that a recent email
has been sent to the user about security best practices. The page has a
confirmation button that can't be skipped. Trying to bypass by directly
accessing other routes does not work too so we must raise an ActionNeeded.