[societegenerale/sgpe] handle errors on subscriptions url
When reaching the subscriptions, the user sometimes reaches
errors; those are similar to some errors reached in the
accounts pages. We thus reuse a similar method to handle
reached errors.
Unencountered errors will have to be added manually later.

Not tested : sca prevents us to really test,
and our sgpe creds are down.